requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
supabase==2.4.6
python-dotenv==1.0.1
